% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsw_county_population.R
\docType{data}
\name{nsw_county_population}
\alias{nsw_county_population}
\title{Historical NSW census data}
\format{A data frame with 47675 rows and 8 variables:
\describe{
  \item{year}{The census year of the data}
  \item{area_type}{Is the area a county or a pastoral district?}
  ...
}}
\source{
\url{http://hccda.ada.edu.au/}
}
\usage{
nsw_county_population
}
\description{
A dataset containing the population of historical counties in NSW
}
\keyword{datasets}
